refactor(callout): move callout client from standalone crate into core
Move the callout client into `core/src/callout/` so it becomes a
first-class framework primitive available to any crate that depends
on `praxis-core`, including custom proxy builds.

The standalone `praxis-proxy-callout` crate is removed; its source
files move into `core/src/callout/` with only import-path adjustments
(no behavioral changes).

Assisted by Opus 4.6
