--- callout/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "praxis-proxy-callout"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-description = "Reusable HTTP callout client for Praxis filters"
-license.workspace = true
-repository.workspace = true
-readme = "../README.md"
-publish = true
-
-[lib]
-name = "praxis_callout"
-
-[lints]
-workspace = true
-
-[dependencies]
-reqwest = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true, features = ["time"] }
-tracing = { workspace = true }
-
-[dev-dependencies]
-tokio = { workspace = true, features = ["macros", "rt-multi-thread", "time"] }
-wiremock = { workspace = true }
