+++ core/src/callout/circuit.rs
