+++ core/src/lib.rs
@@ -5,6 +5,8 @@
 
 //! Core configuration, error types, and server factory for Praxis.
 
+/// Reusable HTTP callout client with circuit breaking and loop prevention.
+pub mod callout;
 /// YAML configuration parsing and validation.
 pub mod config;
 /// Upstream connection options and endpoint types.
